check format with parquet
