+++ be/src/format_v2/table/iceberg_reader.h
@@ -49,8 +49,6 @@ namespace doris::format::iceberg {
 class IcebergTableReader : public format::TableReader {
 public:
     ~IcebergTableReader() override = default;
-    static Status validate_variant_file_mappings(
-            FileFormat format, const std::vector<format::ColumnMapping>& mappings);
     Status init(format::TableReadOptions&& options) override {
         RETURN_IF_ERROR(format::TableReader::init(std::move(options)));
         _mapper_options.mode = format::TableColumnMappingMode::BY_FIELD_ID;
@@ -70,8 +68,6 @@ public:
     }
 
 protected:
-    Status validate_file_mapping(const format::TableColumnMapper& mapper) const override;
-
     void configure_mapper_options(format::TableColumnMapperOptions* options) const override {
         options->enable_row_lineage_virtual_columns = true;
         options->allow_idless_complex_wrapper_projection =
